Tabs
Tabs are perfect for single page web applications, or for web pages capable of displaying different subjects:
Create buttons to open specific tab content. All <div> elements with class="tabcontent" are hidden by default (with CSS & JS). When the user clicks on a button - it will open the tab content that "matches" this button.